requirements.txt
mitmproxy>=7.0.0
requests>=2.25.0
scikit-learn>=0.24.0
pandas>=1.2.0
xgboost>=1.3.0
numpy>=1.19.0
flask>=2.0.0
joblib>=1.0.0
python-Levenshtein>=0.12.0
